split tests up so they can work in parallel
