+++ playwright-e2e/tests/eligibility-tests/eligibility-verify-applicant-UK-resident.spec.ts
@@ -1,0 +1,40 @@
+import { expect, test } from '../../fixtures/fixtures';
+import { Eligibility } from '../../pages/eligibility-pages';
+import { urlConfig } from '../../utils/urls';
+
+const attachTestInfo = async (testInfo, data) => {
+  await testInfo.attach('accessibility-scan-results', {
+    body: JSON.stringify(data, null, 2),
+    contentType: 'application/json',
+  });
+};
+
+test.beforeEach(async ({ page }) => {
+  await page.goto(urlConfig.citizenStartUrl);
+});
+
+test(
+  'Verify you cannot apply to adopt multiple children unless you have a permanent home in the UK, Channel Islands or Isle of Man.',
+  { tag: ['@eligibility', '@citizen', '@accessibility'] },
+  async ({ page, makeAxeBuilder }, testInfo) => {
+    const eligibility = new Eligibility(page);
+    await eligibility.notUKResident(eligibility.applyMoreThanOneChildYes);
+
+    const accessibilityScanResults = await makeAxeBuilder().analyze();
+    await attachTestInfo(testInfo, accessibilityScanResults);
+    expect(accessibilityScanResults.violations).toEqual([]);
+  }
+);
+
+test(
+  'Verify you cannot apply to adopt a child unless you have a permanent home in the UK, Channel Islands or Isle of Man.',
+  { tag: ['@eligibility', '@citizen', '@accessibility'] },
+  async ({ page, makeAxeBuilder }, testInfo) => {
+    const eligibility = new Eligibility(page);
+    await eligibility.notUKResident(eligibility.applyMoreThanOneChildNo);
+
+    const accessibilityScanResults = await makeAxeBuilder().analyze();
+    await attachTestInfo(testInfo, accessibilityScanResults);
+    expect(accessibilityScanResults.violations).toEqual([]);
+  }
+);
